discogs/api/requirements.txt
flask
requests
python3-discogs-client
python-dotenv